requirements.txt
anyio==3.6.1
asgiref==3.5.2
autopep8==1.6.0
backports.cached-property==1.0.1
certifi==2022.5.18.1
click==8.1.3
distlib==0.3.4
fastapi==0.78.0
filelock==3.7.1
graphql-core==3.2.1
h11==0.12.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
pipenv==2022.6.7
platformdirs==2.5.2
prisma==0.6.6
pycodestyle==2.8.0
pydantic==1.9.1
Pygments==2.12.0
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
strawberry-graphql==0.114.1
toml==0.10.2
typing_extensions==4.2.0
uvicorn==0.17.6
uvloop==0.16.0
virtualenv==20.14.1
virtualenv-clone==0.5.7
watchgod==0.8.2
websockets==10.3
